+++ diesel/src/sqlite/types/date_and_time/mod.rs
@@ -9,7 +9,12 @@ use sql_types;
 #[cfg(feature = "chrono")]
 mod chrono;
 
-impl FromSql<sql_types::Date, Sqlite> for String {
+/// The returned pointer is *only* valid for the lifetime to the argument of
+/// `from_sql`. This impl is intended for uses where you want to write a new
+/// impl in terms of `String`, but don't want to allocate. We have to return a
+/// raw pointer instead of a reference with a lifetime due to the structure of
+/// `FromSql`
+impl FromSql<sql_types::Date, Sqlite> for *const str {
     fn from_sql(value: Option<&SqliteValue>) -> deserialize::Result<Self> {
         FromSql::<sql_types::Text, Sqlite>::from_sql(value)
     }
@@ -27,7 +32,12 @@ impl ToSql<sql_types::Date, Sqlite> for String {
     }
 }
 
-impl FromSql<sql_types::Time, Sqlite> for String {
+/// The returned pointer is *only* valid for the lifetime to the argument of
+/// `from_sql`. This impl is intended for uses where you want to write a new
+/// impl in terms of `String`, but don't want to allocate. We have to return a
+/// raw pointer instead of a reference with a lifetime due to the structure of
+/// `FromSql`
+impl FromSql<sql_types::Time, Sqlite> for *const str {
     fn from_sql(value: Option<&SqliteValue>) -> deserialize::Result<Self> {
         FromSql::<sql_types::Text, Sqlite>::from_sql(value)
     }
@@ -45,7 +55,12 @@ impl ToSql<sql_types::Time, Sqlite> for String {
     }
 }
 
-impl FromSql<sql_types::Timestamp, Sqlite> for String {
+/// The returned pointer is *only* valid for the lifetime to the argument of
+/// `from_sql`. This impl is intended for uses where you want to write a new
+/// impl in terms of `String`, but don't want to allocate. We have to return a
+/// raw pointer instead of a reference with a lifetime due to the structure of
+/// `FromSql`
+impl FromSql<sql_types::Timestamp, Sqlite> for *const str {
     fn from_sql(value: Option<&SqliteValue>) -> deserialize::Result<Self> {
         FromSql::<sql_types::Text, Sqlite>::from_sql(value)
     }